requirements.txt
aiofiles
aiohttp
asyncpg
discord.py
fastapi
jaconv
jinja2
PyNaCl
python-dotenv
python-multipart
uvicorn
rich
loguru
aioconsole
pydantic
ffmpeg
romkan2
emoji
uuid